Conformance: UTxO plumbing and translation
