+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway/Base.hs
@@ -62,7 +62,6 @@ import Cardano.Ledger.Conway.Rules (
   spoAccepted,
   spoAcceptedRatio,
  )
-import Cardano.Ledger.Conway.Tx (AlonzoTx)
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Keys (KeyRole (..))
@@ -71,7 +70,6 @@ import Constrained
 import Constrained.Base (fromList_)
 import Data.Bifunctor (Bifunctor (..))
 import Data.Foldable (Foldable (..))
-import qualified Data.List.NonEmpty as NE
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Ratio ((%))
@@ -83,7 +81,6 @@ import qualified Lib as Agda
 import Test.Cardano.Ledger.Common (Arbitrary (..))
 import Test.Cardano.Ledger.Conformance (
   ExecSpecRule (..),
-  OpaqueErrorString (..),
   SpecTranslate (..),
   computationResultToEither,
   runSpecTransM,
@@ -99,62 +96,10 @@ import Test.Cardano.Ledger.Constrained.Conway (
   epochSignalSpec,
   epochStateSpec,
   newEpochStateSpec,
-  utxoEnvSpec,
-  utxoStateSpec,
-  utxoTxSpec,
  )
 import Test.Cardano.Ledger.Constrained.Conway.Instances ()
 import Test.Cardano.Ledger.Conway.Arbitrary ()
 import Test.Cardano.Ledger.Imp.Common hiding (arbitrary, forAll, prop, var)
-
-instance
-  forall fn.
-  IsConwayUniv fn =>
-  ExecSpecRule fn "UTXO" Conway
-  where
-  environmentSpec _ = utxoEnvSpec
-
-  stateSpec _ = utxoStateSpec
-
-  signalSpec _ env st =
-    utxoTxSpec env st
-      <> constrained disableInlineDatums
-    where
-      disableInlineDatums :: Term fn (AlonzoTx Conway) -> Pred fn
-      disableInlineDatums tx = match @fn tx $ \txBody _ _ _ ->
-        match txBody $
-          \_ctbSpendInputs
-           _ctbCollateralInputs
-           _ctbReferenceInputs
-           ctbOutputs
-           _ctbCollateralReturn
-           _ctbTotalCollateral
-           _ctbCerts
-           _ctbWithdrawals
-           _ctbTxfee
-           _ctbVldt
-           _ctbReqSignerHashes
-           _ctbMint
-           _ctbScriptIntegrityHash
-           _ctbAdHash
-           _ctbTxNetworkId
-           _ctbVotingProcedures
-           _ctbProposalProcedures
-           _ctbCurrentTreasuryValue
-           _ctbTreasuryDonation ->
-              match ctbOutputs $
-                \outs -> forAll' outs $
-                  \txOut _ -> match txOut $
-                    \_ _ dat _ ->
-                      (caseOn dat)
-                        (branch $ \_ -> True)
-                        (branch $ \_ -> True)
-                        (branch $ \_ -> False)
-
-  runAgdaRule env st sig =
-    first (\e -> OpaqueErrorString (T.unpack e) NE.:| [])
-      . computationResultToEither
-      $ Agda.utxoStep env st sig
 
 data ConwayCertExecContext era = ConwayCertExecContext
   { ccecWithdrawals :: !(Map (Network, Credential 'Staking (EraCrypto era)) Coin)
