+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway/Utxo.hs
@@ -1,0 +1,72 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Utxo () where
+
+import Cardano.Ledger.Alonzo.Tx (AlonzoTx)
+import Cardano.Ledger.Conway (Conway)
+import Constrained
+import Data.Bifunctor (first)
+import qualified Data.List.NonEmpty as NE
+import qualified Data.Text as T
+import qualified Lib as Agda
+import Test.Cardano.Ledger.Conformance
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Base ()
+import Test.Cardano.Ledger.Constrained.Conway
+import Test.Cardano.Ledger.Conway.Arbitrary ()
+
+instance
+  forall fn.
+  IsConwayUniv fn =>
+  ExecSpecRule fn "UTXO" Conway
+  where
+  environmentSpec _ = utxoEnvSpec
+
+  stateSpec _ = utxoStateSpec
+
+  signalSpec _ env st =
+    utxoTxSpec env st
+      <> constrained disableInlineDatums
+    where
+      disableInlineDatums :: Term fn (AlonzoTx Conway) -> Pred fn
+      disableInlineDatums tx = match @fn tx $ \txBody _ _ _ ->
+        match txBody $
+          \_ctbSpendInputs
+           _ctbCollateralInputs
+           _ctbReferenceInputs
+           ctbOutputs
+           _ctbCollateralReturn
+           _ctbTotalCollateral
+           _ctbCerts
+           _ctbWithdrawals
+           _ctbTxfee
+           _ctbVldt
+           _ctbReqSignerHashes
+           _ctbMint
+           _ctbScriptIntegrityHash
+           _ctbAdHash
+           _ctbTxNetworkId
+           _ctbVotingProcedures
+           _ctbProposalProcedures
+           _ctbCurrentTreasuryValue
+           _ctbTreasuryDonation ->
+              match ctbOutputs $
+                \outs -> forAll' outs $
+                  \txOut _ -> match txOut $
+                    \_ _ dat _ ->
+                      (caseOn dat)
+                        (branch $ \_ -> True)
+                        (branch $ \_ -> True)
+                        (branch $ \_ -> False)
+
+  runAgdaRule env st sig =
+    first (\e -> OpaqueErrorString (T.unpack e) NE.:| [])
+      . computationResultToEither
+      $ Agda.utxoStep env st sig
