+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Utxo.hs
@@ -1,0 +1,50 @@
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE NumericUnderscores #-}
+{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Utxo () where
+
+import Cardano.Ledger.Coin
+import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Shelley.LedgerState (UTxOState (..))
+import Cardano.Ledger.Shelley.Rules (UtxoEnv (..))
+import Data.Functor.Identity (Identity)
+import qualified Lib as Agda
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base ()
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Core
+
+instance
+  ( SpecRep (PParams era) ~ Agda.PParams
+  , SpecTranslate ctx (PParamsHKD Identity era)
+  ) =>
+  SpecTranslate ctx (UtxoEnv era)
+  where
+  type SpecRep (UtxoEnv era) = Agda.UTxOEnv
+
+  toSpecRep x =
+    Agda.MkUTxOEnv
+      <$> toSpecRep (ueSlot x)
+      <*> toSpecRep (uePParams x)
+      <*> toSpecRep (Coin 10_000_000) -- TODO: Fix generating types
+
+instance
+  ( SpecTranslate ctx (TxOut era)
+  , SpecRep (TxOut era) ~ Agda.TxOut
+  ) =>
+  SpecTranslate ctx (UTxOState era)
+  where
+  type SpecRep (UTxOState era) = Agda.UTxOState
+
+  toSpecRep x =
+    Agda.MkUTxOState
+      <$> toSpecRep (utxosUtxo x)
+      <*> toSpecRep (utxosFees x)
